+++ CHANGELOG.rst
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 ====================
+ Unreleased
+====================
+
+Added
+_____
+
+* Cloud only: New regions: us-tacoma-1, us-chicago-1, eu-dcc-madrid-1.
+
+====================
  5.3.5 - 2022-08-09
 ====================